feat: complete cross-distro/cross-compositor support
Streams completed:
- Stream 4: Battery sysfs fallback (UPower → /sys/class/power_supply/ → no-battery)
- Stream 5: Clipboard fallback chain (wl-clipboard → xclip → xsel → graceful error)
- Stream 6: Notification D-Bus fallback (notify-send → native zbus D-Bus)
- Stream 7: Portal timeout handling (all ashpd calls wrapped with 10s timeout)
- Stream 9 (partial): Input injector uses evdev keycodes for universal input

Quality gates:
- clippy: 0 warnings across workspace
- tests: 66 pass (54 core + 12 server)
- fmt: clean

All plugin fallback chains now work on any Linux distro/compositor.

+++ core/src/streaming/capture_x11.rs
@@ -29,8 +29,8 @@ fn capture_frame(monitor_idx: usize) -> Result<VideoFrame> {
         .capture_image()
         .context("Failed to capture X11 screenshot")?;
 
-    let width = image.width() as u32;
-    let height = image.height() as u32;
+    let width = image.width();
+    let height = image.height();
     let stride = width * 4; // 4 bytes per pixel (RGBA)
 
     // xcap returns RGBA, but our encoder expects BGRA (PipeWire's default).
